Add icons to create node dialog
Extract IconCache into a shared module so both NetworkView and
NodeSelectionDialog can display node icons. The dialog now shows
actual PNG icons from the libraries folder, with fallback to
procedural category-based icons when PNG files are not available.
